rag-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-generativeai==0.8.5
pydantic==2.9.2
